DialogFlow/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
uuid==1.30
python-jose==3.3.0
httpx==0.25.1
toml==0.10.2
jinja2==3.1.2
spacy==3.7.2
pytz==2023.3 